pyproject.toml
[tool.poetry]
name = "thps"
version = "0.1.0"
description = ""
authors = ["David Ambrozio <davidambroziodev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
selenium = "^4.29.0"
fastapi = "^0.115.11"
webdriver-manager = "^4.0.2"
python-dotenv = "^1.1.0"
openai = "^1.74.0"
tk = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
